[SMTChecker] Split SMTChecker into SMTEncoder and BMC
