+++ test/libsolidity/smtCheckerTests/complex/slither/const_state_variables.sol
@@ -58,3 +58,4 @@ contract MyConc{
 // Warning: (834-839): Assertion checker does not yet support the type of this literal (literal_string "abc").
 // Warning: (874-879): Underflow (resulting value less than 0) happens here.
 // Warning: (874-879): Overflow (resulting value larger than 2**256 - 1) happens here.
+// Warning: (985-1002): Overflow (resulting value larger than 2**256 - 1) happens here.
